#157 Performance optimization. A workaround to cache stores (do not load stores for each HTTP request).

+++ src/Libraries/Nop.Services/Stores/StoreService.cs
@@ -32,12 +32,12 @@ namespace Nop.Services.Stores
         private const string STORES_PATTERN_KEY = "Nop.stores.";
 
         #endregion
-        
+
         #region Fields
-        
+
         private readonly IRepository<Store> _storeRepository;
         private readonly IEventPublisher _eventPublisher;
-        private readonly ICacheManager _cacheManager;
+        private readonly IStaticCacheManager _staticCacheManager;
 
         #endregion
 
@@ -46,14 +46,14 @@ namespace Nop.Services.Stores
         /// <summary>
         /// Ctor
         /// </summary>
-        /// <param name="cacheManager">Cache manager</param>
+        /// <param name="staticCacheManager">Static cache manager</param>
         /// <param name="storeRepository">Store repository</param>
         /// <param name="eventPublisher">Event published</param>
-        public StoreService(ICacheManager cacheManager,
+        public StoreService(IStaticCacheManager staticCacheManager,
             IRepository<Store> storeRepository,
             IEventPublisher eventPublisher)
         {
-            this._cacheManager = cacheManager;
+            this._staticCacheManager = staticCacheManager;
             this._storeRepository = storeRepository;
             this._eventPublisher = eventPublisher;
         }
@@ -71,13 +71,16 @@ namespace Nop.Services.Stores
             if (store == null)
                 throw new ArgumentNullException(nameof(store));
 
+            if (store is StoreForCaching)
+                throw new ArgumentException("Cacheable entities are not supported by Entity Framework");
+
             var allStores = GetAllStores();
             if (allStores.Count == 1)
                 throw new Exception("You cannot delete the only configured store");
 
             _storeRepository.Delete(store);
 
-            _cacheManager.RemoveByPattern(STORES_PATTERN_KEY);
+            _staticCacheManager.RemoveByPattern(STORES_PATTERN_KEY);
 
             //event notification
             _eventPublisher.EntityDeleted(store);
@@ -86,33 +89,67 @@ namespace Nop.Services.Stores
         /// <summary>
         /// Gets all stores
         /// </summary>
+        /// <param name="loadCacheableCopy">A value indicating whether to load a copy that could be cached (workaround until Entity Framework supports 2-level caching)</param>
         /// <returns>Stores</returns>
-        public virtual IList<Store> GetAllStores()
+        public virtual IList<Store> GetAllStores(bool loadCacheableCopy = true)
         {
-            string key = STORES_ALL_KEY;
-            return _cacheManager.Get(key, () =>
+            Func<IList<Store>> loadStoresFunc = () =>
             {
                 var query = from s in _storeRepository.Table
-                            orderby s.DisplayOrder, s.Id
-                            select s;
-                var stores = query.ToList();
-                return stores;
-            });
+                    orderby s.DisplayOrder, s.Id
+                    select s;
+                return query.ToList();
+            };
+
+            if (loadCacheableCopy)
+            {
+                //cacheable copy
+                return _staticCacheManager.Get(STORES_ALL_KEY, () =>
+                {
+                    var result = new List<Store>();
+                    foreach (var store in loadStoresFunc())
+                        result.Add(new StoreForCaching(store));
+                    return result;
+                });
+            }
+            else
+            {
+                return loadStoresFunc();
+            }
         }
 
         /// <summary>
         /// Gets a store 
         /// </summary>
         /// <param name="storeId">Store identifier</param>
+        /// <param name="loadCacheableCopy">A value indicating whether to load a copy that could be cached (workaround until Entity Framework supports 2-level caching)</param>
         /// <returns>Store</returns>
-        public virtual Store GetStoreById(int storeId)
+        public virtual Store GetStoreById(int storeId, bool loadCacheableCopy = true)
         {
             if (storeId == 0)
                 return null;
-            
-            string key = string.Format(STORES_BY_ID_KEY, storeId);
-            return _cacheManager.Get(key, () => _storeRepository.GetById(storeId));
-        }
+
+            Func<Store> loadStoreFunc = () =>
+            {
+                return _storeRepository.GetById(storeId);
+            };
+
+            if (loadCacheableCopy)
+            {
+                //cacheable copy
+                return _staticCacheManager.Get(string.Format(STORES_BY_ID_KEY, storeId), () =>
+                {
+                    var store = loadStoreFunc();
+                    if (store == null)
+                        return null;
+                    return new StoreForCaching(store);
+                });
+            }
+            else
+            {
+                return loadStoreFunc();
+            }
+        } 
 
         /// <summary>
         /// Inserts a store
@@ -122,10 +159,13 @@ namespace Nop.Services.Stores
         {
             if (store == null)
                 throw new ArgumentNullException(nameof(store));
+            
+            if (store is StoreForCaching)
+                throw  new ArgumentException("Cacheable entities are not supported by Entity Framework");
 
             _storeRepository.Insert(store);
 
-            _cacheManager.RemoveByPattern(STORES_PATTERN_KEY);
+            _staticCacheManager.RemoveByPattern(STORES_PATTERN_KEY);
 
             //event notification
             _eventPublisher.EntityInserted(store);
@@ -140,9 +180,12 @@ namespace Nop.Services.Stores
             if (store == null)
                 throw new ArgumentNullException(nameof(store));
 
+            if (store is StoreForCaching)
+                throw new ArgumentException("Cacheable entities are not supported by Entity Framework");
+
             _storeRepository.Update(store);
 
-            _cacheManager.RemoveByPattern(STORES_PATTERN_KEY);
+            _staticCacheManager.RemoveByPattern(STORES_PATTERN_KEY);
 
             //event notification
             _eventPublisher.EntityUpdated(store);
